+++ packages/babel-plugin-jest-hoist/src/index.ts
@@ -154,7 +154,7 @@ FUNCTIONS.deepUnmock = args => args.length === 1 && args[0].isStringLiteral();
 FUNCTIONS.disableAutomock = FUNCTIONS.enableAutomock = args =>
   args.length === 0;
 
-export = () => {
+export default () => {
   const shouldHoistExpression = (expr: NodePath): boolean => {
     if (!expr.isCallExpression()) {
       return false;